Remove change_loader_path_for_layout utility.
We really shouldn't encourage hardcoding rpaths to paclets in the
layout...this mechanism is fragile and will break if the dependency
paclet ever gets updated and installed into a new location.
For libcurl, AudioFileStreamTools and FFmpegTools should just
be rewritten to use CURLLink. For libmp3lame, ... I don't have a good
solution for AudioFileStreamTools and SoundFileTools, but currently
the install name for libmp3lame.dylib on cvs is already hardcoded to
be @executable_path/../SystemFiles/Links/MP3Tools/etc, so it is moot.
I don't know of any other paclets currently relying on this utility.